requirements.txt
langchain-openai==0.3.0
fastapi==0.115.6
uvicorn==0.34.0
httpx==0.28.1
pydantic==2.10.5
pydantic-core==2.27.2
python-dotenv==1.0.1